Found it!
This one is fun. So the List I made is being implicitly copied, but its destructor `delete`s its ListItems when one instance is destructed, leaving the other with invalid pointers. To fix this I should

a) Write a proper copy constructor. This wastes loads of memory
b) Implement const methods to prevent this from happening